restructure sbk-chart python files
